Add new contact protocols and implementation were added
ContactTableVC was modified to navigate from Table to add new contact with presenter.
tableContactPresenter was modified to use router to navigate to add new contact view.
table router was modified to add Add router, to navigate from table to add new contact

navigationToAddNewContact functions was added to presenter protocol and to router protocol