Temporal Contact data was implemented Detail contact protocols were created with some methods
Detail implementation were created with protocols implementation

DetailVC was implemented with contact to be show when user touch into table cells.

ContactTableVC was modified to add navigationDetail with contact detail (dummy contact at the moment)

TableContactPresenter was modified to navigate from table cell to detail, passing a contact to detail.

TableContactRouter was modified to pass a contact detail when user touch a table cell